Move modeset test C into readable fixture files
